PlantMeteo

Documentation for PlantMeteo.

DataFrames.DataFrame

PlantMeteo.AbstractAtmosphere

PlantMeteo.Atmosphere

PlantMeteo.Constants

PlantMeteo.Weather

PlantMeteo.air_density

PlantMeteo.atmosphere_emissivity

PlantMeteo.e_sat

PlantMeteo.e_sat_slope

PlantMeteo.latent_heat_vaporization

PlantMeteo.psychrometer_constant

PlantMeteo.rh_from_e

PlantMeteo.rh_from_vpd

PlantMeteo.vapor_pressure

PlantMeteo.vpd

PlantMeteo.γ_star

DataFrames.DataFrame — Method

DataFrame(data::Weather)

Transform a Weather type into a DataFrame.

See also Weather to make the reverse.

PlantMeteo.AbstractAtmosphere — Type

Abstract atmospheric conditions type. The suptypes of AbstractAtmosphere should describe the atmospheric conditions for one time-step only, see e.g. Atmosphere

PlantMeteo.Atmosphere — Type

Atmosphere structure to hold all values related to the meteorology / atmosphere.

Arguments

date = Dates.now(): the date of the record.

duration = 1.0 (seconds): the duration of the time-step.

T (°C): air temperature

Wind (m s-1): wind speed

P = 101.325 (kPa): air pressure. The default value is at 1 atm, i.e. the mean sea-level

atmospheric pressure on Earth.

Rh = rh_from_vpd(VPD,eₛ) (0-1): relative humidity

Precipitations=0.0 (mm): precipitations from atmosphere (i.e. rain, snow, hail, etc.)

Cₐ (ppm): air CO₂ concentration

e = vapor_pressure(T,Rh) (kPa): vapor pressure

eₛ = e_sat(T) (kPa): saturated vapor pressure

VPD = eₛ - e (kPa): vapor pressure deficit

ρ = air_density(T, P, constants.Rd, constants.K₀) (kg m-3): air density

λ = latent_heat_vaporization(T, constants.λ₀) (J kg-1): latent heat of vaporization

γ = psychrometer_constant(P, λ, constants.Cₚ, constants.ε) (kPa K−1): psychrometer "constant"

ε = atmosphere_emissivity(T,e,constants.K₀) (0-1): atmosphere emissivity

Δ = e_sat_slope(meteo.T) (0-1): slope of the saturation vapor pressure at air temperature

clearness::A = 9999.9 (0-1): Sky clearness

Ri_SW_f::A = 9999.9 (W m-2): Incoming short wave radiation flux

Ri_PAR_f::A = 9999.9 (W m-2): Incoming PAR flux

Ri_NIR_f::A = 9999.9 (W m-2): Incoming NIR flux

Ri_TIR_f::A = 9999.9 (W m-2): Incoming TIR flux

Ri_custom_f::A = 9999.9 (W m-2): Incoming radiation flux for a custom waveband

Notes

The structure can be built using only T, Rh, Wind and P. All other variables are optional and either let at their default value or automatically computed using the functions given in Arguments.

Examples

Atmosphere(T = 20.0, Wind = 1.0, P = 101.3, Rh = 0.65)

PlantMeteo.Constants — Type

Physical constants

The definition and default values are:

K₀ = -273.15: absolute zero (°C)

R = 8.314: universal gas constant ($J\ mol^{-1}\ K^{-1}$).

Rd = 287.0586: gas constant of dry air ($J\ kg^{-1}\ K^{-1}$).

Dₕ₀ = 21.5e-6: molecular diffusivity for heat at base temperature, applied in the integrated form of the Fick’s Law of diffusion ($m^2\ s^{-1}$). See eq. 3.10 from Monteith and Unsworth (2013).

Cₚ = 1013.0: Specific heat of air at constant pressure ($J\ K^{-1}\ kg^{-1}$), also known as efficiency of impaction of particles. See Allen et al. (1998), or Monteith and Unsworth (2013). NB: bigleaf R package uses 1004.834 intead.

ε = 0.622: ratio of molecular weights of water vapor and air. See Monteith and Unsworth (2013).

λ₀ = 2.501: latent heat of vaporization for water at 0 degree ($J\ kg^{-1}$).

σ = 5.670373e-08 Stefan-Boltzmann constant in ($W\ m^{-2}\ K^{-4}$).

Gbₕ_to_Gbₕ₂ₒ = 1.075: conversion coefficient from conductance to heat to conductance to water vapor.

Gsc_to_Gsw = 1.57: conversion coefficient from stomatal conductance to CO₂ to conductance to water vapor.

Gbc_to_Gbₕ = 1.32: conversion coefficient from boundary layer conductance to CO₂ to heat.

Mₕ₂ₒ = 18.0e-3 (kg mol-1): Molar mass for water.

References

Allen, Richard G., Luis S. Pereira, Dirk Raes, et Martin J Fao Smith. 1998. « Crop evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56 » 300 (9): D05109.

Monteith, John, et Mike Unsworth. 2013. Principles of environmental physics: plants, animals, and the atmosphere. Academic Press.

PlantMeteo.Weather — Type

Weather(D <: AbstractArray{<:AbstractAtmosphere}[, S])
Weather(df::DataFrame[, mt])

Defines the weather, i.e. the local conditions of the Atmosphere for one or more time-steps. Each time-step is described using the Atmosphere structure.

The simplest way to instantiate a Weather is to use a DataFrame as input.

The DataFrame should be formated such as each row is an observation for a given time-step and each column is a variable. The column names should match exactly the field names of the Atmosphere structure, i.e.:

fieldnames(Atmosphere)

See also

the Atmosphere structure

the read_weather function to read Archimed-formatted meteorology data.

Examples

# Example of weather data defined by hand (cumbersome):
w = Weather(
    [
        Atmosphere(T = 20.0, Wind = 1.0, P = 101.3, Rh = 0.65),
        Atmosphere(T = 23.0, Wind = 1.5, P = 101.3, Rh = 0.60),
        Atmosphere(T = 25.0, Wind = 3.0, P = 101.3, Rh = 0.55)
    ],
    (
        site = "Test site",
        important_metadata = "this is important and will be attached to our weather data"
    )
)

# Example using a DataFrame, that you would usually import from a file:
using CSV, DataFrames
file = joinpath(dirname(dirname(pathof(PlantBiophysics))),"test","inputs","meteo.csv")
df = CSV.read(file, DataFrame; header=5, datarow = 6)
# Select and rename the variables:
select!(df, :date, :VPD, :temperature => :T, :relativeHumidity => :Rh, :wind => :Wind, :atmosphereCO2_ppm => :Cₐ)
df[!,:duration] .= 1800 # Add the time-step duration, 30min

# Make the weather, and add some metadata:
Weather(df, (site = "Aquiares", file = file))

PlantMeteo.air_density — Method

air_density(Tₐ, P)
air_density(Tₐ, P, Rd, K₀)

ρ, the air density (kg m-3).

Arguments

Tₐ (Celsius degree): air temperature

P (kPa): air pressure

Rd (J kg-1 K-1): gas constant of dry air (see Foken p. 245, or R bigleaf package).

K₀ (Celsius degree): temperature in Celsius degree at 0 Kelvin

Note

Rd and K₀ are Taken from Constants if not provided.

References

Foken, T, 2008: Micrometeorology. Springer, Berlin, Germany.

PlantMeteo.atmosphere_emissivity — Method

atmosphere_emissivity(Tₐ,eₐ)

Emissivity of the atmoshpere at a given temperature and vapor pressure.

Arguments

Tₐ (°C): air temperature

eₐ (kPa): air vapor pressure

K₀ (°C): absolute zero

Examples

Tₐ = 20.0
VPD = 1.5
atmosphere_emissivity(Tₐ, vapor_pressure(Tₐ,VPD))

References

Leuning, R., F. M. Kelliher, DGG de Pury, et E.-D. SCHULZE. 1995. Leaf nitrogen, photosynthesis, conductance and transpiration: scaling from leaves to canopies ». Plant, Cell & Environment 18 (10): 1183‑1200.

PlantMeteo.e_sat — Method

e_sat(T)

Saturated water vapour pressure (es, in kPa) at given temperature T (°C). See Jones (1992) p. 110 for the equation.

PlantMeteo.e_sat_slope — Method

e_sat_slope(T)

Slope of the vapor pressure saturation curve at a given temperature T (°C).

PlantMeteo.latent_heat_vaporization — Method

latent_heat_vaporization(Tₐ,λ₀)
latent_heat_vaporization(Tₐ)

λ, the latent heat of vaporization for water (J kg-1).

Arguments

Tₐ (°C): air temperature

λ₀: latent heat of vaporization for water at 0 degree Celsius. Taken from Constants().λ₀

if not provided (see Constants).

PlantMeteo.psychrometer_constant — Method

psychrometer_constant(P, λ, Cₚ, ε)
psychrometer_constant(P, λ)

γ, the psychrometer constant, also called psychrometric constant (kPa K−1). See Monteith and Unsworth (2013), p. 222.

Arguments

P (kPa): air pressure

λ ($J\ kg^{-1}$): latent heat of vaporization for water (see latent_heat_vaporization)

Cₚ (J kg-1 K-1): specific heat of air at constant pressure ($J\ K^{-1}\ kg^{-1}$)

ε (Celsius degree): temperature in Celsius degree at 0 Kelvin

Note

Cₚ, ε and λ₀ are taken from Constants if not provided.

Tₐ = 20.0

λ = latent_heat_vaporization(Tₐ, λ₀)
psychrometer_constant(100.0, λ)

References

Monteith, John L., et Mike H. Unsworth. 2013. « Chapter 13 - Steady-State Heat Balance: (i) Water Surfaces, Soil, and Vegetation ». In Principles of Environmental Physics (Fourth Edition), edited by John L. Monteith et Mike H. Unsworth, 217‑47. Boston: Academic Press.

PlantMeteo.rh_from_e — Method

rh_from_e(VPD,eₛ)

Conversion between e (kPa) and rh (0-1).

Examples

rh_from_e(1.5,25.0)

PlantMeteo.rh_from_vpd — Method

rh_from_vpd(VPD,eₛ)

Conversion between VPD and rh.

Examples

eₛ = e_sat(Tₐ)
rh_from_vpd(1.5,eₛ)

PlantMeteo.vapor_pressure — Method

vapor_pressure(Tₐ, rh)

Vapor pressure (kPa) at given temperature (°C) and relative hunidity (0-1).

PlantMeteo.vpd — Method

vpd(VPD,eₛ)

Compute vapor pressure deficit (kPa) from the air relative humidity (0-1) and temperature (°C).

The computation simply uses vpd = eₛ - e.

Examples

vpd(0.4,25.0)

PlantMeteo.γ_star — Method

γstar(γ, ash, a_s, rbv, Rsᵥ, Rbₕ)

γ∗, the apparent value of psychrometer constant (kPa K−1).

Arguments

γ (kPa K−1): psychrometer constant

aₛₕ (1,2): number of faces exchanging heat fluxes (see Schymanski et al., 2017)

aₛᵥ (1,2): number of faces exchanging water fluxes (see Schymanski et al., 2017)

Rbᵥ (s m-1): boundary layer resistance to water vapor

Rsᵥ (s m-1): stomatal resistance to water vapor

Rbₕ (s m-1): boundary layer resistance to heat

Note

Using the corrigendum from Schymanski et al. (2017) in here so the definition of latent_heat remains generic.

Not to be confused with Γ_star in FcVB model, which is the CO₂ compensation point.

References

Monteith, John L., et Mike H. Unsworth. 2013. « Chapter 13 - Steady-State Heat Balance: (i) Water Surfaces, Soil, and Vegetation ». In Principles of Environmental Physics (Fourth Edition), edited by John L. Monteith et Mike H. Unsworth, 217‑47. Boston: Academic Press.

Schymanski, Stanislaus J., et Dani Or. 2017. Leaf-Scale Experiments Reveal an Important Omission in the Penman–Monteith Equation ». Hydrology and Earth System Sciences 21 (2): 685‑706. https://doi.org/10.5194/hess-21-685-2017.